+++ packages/client/src/utils/fpsMeter.ts
@@ -1,31 +1,31 @@
 let fpsAcc = 0;
 let fpsTime = 0;
 export let stats = {
-    drawCalls: 0,
-    vertices: 0,
-    triangles: 0,
+    _drawCalls: 0,
+    _vertices: 0,
+    _triangles: 0,
 
-    frameDrawCalls: 0,
-    frameVertices: 0,
-    frameTriangles: 0,
+    _frameDrawCalls: 0,
+    _frameVertices: 0,
+    _frameTriangles: 0,
 
-    fps: 0,
+    _fps: 0,
 };
 
 export const updateStats = (ts: number) => {
-    stats.drawCalls = stats.frameDrawCalls;
-    stats.frameDrawCalls = 0;
+    stats._drawCalls = stats._frameDrawCalls;
+    stats._frameDrawCalls = 0;
 
-    stats.vertices = stats.frameVertices;
-    stats.frameVertices = 0;
+    stats._vertices = stats._frameVertices;
+    stats._frameVertices = 0;
 
-    stats.triangles = stats.frameTriangles;
-    stats.frameTriangles = 0;
+    stats._triangles = stats._frameTriangles;
+    stats._frameTriangles = 0;
 
     ++fpsAcc;
     const div = 1;
     if ((ts - fpsTime) >= div) {
-        stats.fps = fpsAcc / div;
+        stats._fps = fpsAcc / div;
         fpsAcc = 0;
         fpsTime += div;
     }